moved request token outside of apis and fixed review request childprocess
